pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "autorooms"
description = "Zero Config Autoroom Discord Bot"
readme = "README.md"
license = "	Apache-2.0"
requires-python = ">=3.8, <3.12"
authors = [{ name = "Michael Hall", email = "michael@michaelhall.tech" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Typing :: Typed",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
]
dynamic = ["version"]
dependencies = [
    "discord.py>=2.2.0, <3",
    "orjson>=3.5.4",
    "aiodns>=1.1",
    "Brotli",
    "uvloop>=0.17.0; platform_system != 'Windows'",
]

[project.urls]
Homepage = "https://github.com/mikeshardmind/autorooms"

[tool.hatch.version]
path = "autorooms/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/autorooms",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
include = ["/autorooms/*"]


[project.scripts]
discord-autorooms = "autorooms.__main__:main"


[tool.ruff]

line-length = 120
target-version = "py38"
select = [
    "F", "E", "I", "C90", "UP", "YTT", "S", "BLE", "B", "A", "COM", "C4", "DTZ",
    "EM", "ISC", "G", "INP", "PIE", "T20", "Q003", "RSE", "RET", "SIM", "TID", "PTH",
    "ERA", "PD", "PLC", "PLE", "PLR", "PLW", "TRY", "NPY", "RUF"
]
ignore = ["G002", "TRY002", "S101", "PLR2004", "SIM102", "SIM105"]
unfixable = ["C90", "S101", "ERA"]
